fix(AwsVpcPeering): waits pending acceptance status before accepting connection

+++ pkg/kcp/provider/aws/vpcpeering/deleteVpcPeering.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"github.com/kyma-project/cloud-manager/pkg/composed"
 	awsmeta "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/meta"
+	awsutil "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/util"
+	"github.com/kyma-project/cloud-manager/pkg/util"
+	"k8s.io/utils/ptr"
 )
 
 func deleteVpcPeering(ctx context.Context, st composed.State) (error, context.Context) {
@@ -15,18 +18,19 @@ func deleteVpcPeering(ctx context.Context, st composed.State) (error, context.Co
 		return nil, nil
 	}
 
+	if !awsutil.CanDeletePeering(state.vpcPeering) {
+		logger.Info("VpcPeering can't be deleted at this stage",
+			"peeringStatusCode", string(state.vpcPeering.Status.Code),
+			"peeringStatusMessage", ptr.Deref(state.vpcPeering.Status.Message, ""))
+		return nil, nil
+	}
+
 	logger.Info("Deleting VpcPeering")
 
 	err := state.client.DeleteVpcPeeringConnection(ctx, state.vpcPeering.VpcPeeringConnectionId)
 
-	if err != nil {
-
-		if awsmeta.IsNotFound(err) {
-			logger.Info("VpcPeeringConnection not found")
-			return nil, nil
-		}
-
-		return awsmeta.LogErrorAndReturn(err, "Error deleting vpc peering", ctx)
+	if awsmeta.IsErrorRetryable(err) {
+		return composed.StopWithRequeueDelay(util.Timing.T10000ms()), nil
 	}
 
 	logger.Info("VpcPeering deleted")
